enhancement: performance optimizations

+++ src/common/RedoLogRecord.cpp
@@ -30,13 +30,13 @@ namespace OpenLogReplicator {
            " op: " << std::setfill('0') << std::setw(4) << std::hex << redo.opCode <<
            " cls: " << std::dec << redo.cls <<
            " rbl: " << std::dec << redo.rbl <<
-           " seq: " << std::dec << static_cast<uint64_t>(redo.seq) <<
-           " typ: " << std::dec << static_cast<uint64_t>(redo.typ) <<
+           " seq: " << std::dec << static_cast<uint>(redo.seq) <<
+           " typ: " << std::dec << static_cast<uint>(redo.typ) <<
            " conId: " << std::dec << redo.conId <<
            " flgRecord: " << std::dec << redo.flgRecord <<
            " robj: " << std::dec << redo.recordObj <<
            " rdataObj: " << std::dec << redo.recordDataObj <<
-           " nrow: " << std::dec << static_cast<uint64_t>(redo.nRow) <<
+           " nrow: " << std::dec << static_cast<uint>(redo.nRow) <<
            " afn: " << std::dec << redo.afn <<
            " size: " << std::dec << redo.size <<
            " dba: 0x" << std::hex << redo.dba <<
@@ -44,14 +44,14 @@ namespace OpenLogReplicator {
            " obj: " << std::dec << redo.obj <<
            " dataobj: " << std::dec << redo.dataObj <<
            " usn: " << std::dec << redo.usn <<
-           " slt: " << std::dec << static_cast<uint64_t>(redo.slt) <<
-           " flg: " << std::dec << static_cast<uint64_t>(redo.flg) <<
+           " slt: " << std::dec << static_cast<uint>(redo.slt) <<
+           " flg: " << std::dec << static_cast<uint>(redo.flg) <<
            " opc: 0x" << std::hex << redo.opc <<
-           " op: " << std::dec << static_cast<uint64_t>(redo.op) <<
-           " cc: " << std::dec << static_cast<uint64_t>(redo.cc) <<
+           " op: " << std::dec << static_cast<uint>(redo.op) <<
+           " cc: " << std::dec << static_cast<uint>(redo.cc) <<
            " slot: " << std::dec << redo.slot <<
-           " flags: 0x" << std::hex << static_cast<uint64_t>(redo.flags) <<
-           " fb: 0x" << std::hex << static_cast<uint64_t>(redo.fb);
+           " flags: 0x" << std::hex << static_cast<uint>(redo.flags) <<
+           " fb: 0x" << std::hex << static_cast<uint>(redo.fb);
         os << ss.str();
         return os;
     }
